Fixes for re-enabling more MSVC level 4 warnings: content/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Possibly-uninitialized local variable
* Assignment inside conditional
* Nonstandard extension: unnamed struct/union

This also contains a very small number of other cleanups/simplifications to
nearby code.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/377723002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283161 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a2779cc129b92a502c615989a688b05526a47b67

+++ content/child/npapi/plugin_instance.cc
@@ -406,7 +406,8 @@ void PluginInstance::DidReceiveManualResponse(const GURL& url,
                                               uint32 last_modified) {
   DCHECK(load_manually_);
 
-  plugin_data_stream_ = CreateStream(-1, url, mime_type, 0);
+  plugin_data_stream_ =
+      CreateStream(static_cast<unsigned long>(-1), url, mime_type, 0);
   plugin_data_stream_->DidReceiveResponse(mime_type, headers, expected_length,
                                           last_modified, true);
 }
